ncc: new intermediate code

+++ mem.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "mem.h"
+#include "ncc.h"
 
 #define MEMSZ		512
 
@@ -27,17 +27,17 @@ void mem_done(struct mem *mem)
 	memset(mem, 0, sizeof(*mem));
 }
 
-void mem_cut(struct mem *mem, int pos)
+void mem_cut(struct mem *mem, long pos)
 {
 	mem->n = pos < mem->n ? pos : mem->n;
 }
 
-void mem_cpy(struct mem *mem, int off, void *buf, int len)
+void mem_cpy(struct mem *mem, long off, void *buf, long len)
 {
 	memcpy(mem->s + off, buf, len);
 }
 
-void mem_put(struct mem *mem, void *buf, int len)
+void mem_put(struct mem *mem, void *buf, long len)
 {
 	while (mem->n + len + 1 >= mem->sz)
 		mem_extend(mem);
@@ -52,7 +52,7 @@ void mem_putc(struct mem *mem, int c)
 	mem->s[mem->n++] = c;
 }
 
-void mem_putz(struct mem *mem, int sz)
+void mem_putz(struct mem *mem, long sz)
 {
 	while (mem->n + sz + 1 >= mem->sz)
 		mem_extend(mem);
@@ -69,7 +69,7 @@ void *mem_buf(struct mem *mem)
 	return mem->s;
 }
 
-int mem_len(struct mem *mem)
+long mem_len(struct mem *mem)
 {
 	return mem->n;
 }
